+++ optaplanner-core/src/main/java/org/optaplanner/core/api/score/calculator/EasyScoreCalculator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.core.impl.score.director.easy;
+package org.optaplanner.core.api.score.calculator;
 
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
 import org.optaplanner.core.api.score.Score;
