+++ optaplanner-core/src/main/java/org/optaplanner/core/impl/score/director/easy/EasyScoreDirectorFactory.java
@@ -18,6 +18,7 @@ package org.optaplanner.core.impl.score.director.easy;
 
 import org.optaplanner.core.api.domain.solution.PlanningSolution;
 import org.optaplanner.core.api.score.Score;
+import org.optaplanner.core.api.score.calculator.EasyScoreCalculator;
 import org.optaplanner.core.impl.domain.solution.descriptor.SolutionDescriptor;
 import org.optaplanner.core.impl.score.director.AbstractScoreDirectorFactory;
 import org.optaplanner.core.impl.score.director.ScoreDirectorFactory;
