+++ optaplanner-examples/src/main/java/org/optaplanner/examples/nqueens/solver/score/NQueensEasyScoreCalculator.java
@@ -19,7 +19,7 @@ package org.optaplanner.examples.nqueens.solver.score;
 import java.util.List;
 
 import org.optaplanner.core.api.score.buildin.simple.SimpleScore;
-import org.optaplanner.core.impl.score.director.easy.EasyScoreCalculator;
+import org.optaplanner.core.api.score.calculator.EasyScoreCalculator;
 import org.optaplanner.examples.nqueens.domain.NQueens;
 import org.optaplanner.examples.nqueens.domain.Queen;
